Chance price datatype to decimal

+++ server/db/models/product.js
@@ -11,7 +11,7 @@ const Product = db.define('product', {
     type: Sequelize.TEXT
   },
   price: {
-    type: Sequelize.FLOAT
+    type: Sequelize.DECIMAL(10,2)
   },
   quantityAvailable: {
     type: Sequelize.INTEGER,
